fix(simulation): support multi-million entity identities

+++ engine/simulation/ecs/core/include/lux/engine/simulation/ecs/Entity.hpp
@@ -6,7 +6,9 @@
 
 namespace lux::simulation::ecs
 {
-    enum class Entity : std::uint32_t
+    // EnTT divides this identity evenly between the live-entity index and
+    // generation, preserving exact reuse checks beyond two million entities.
+    enum class Entity : std::uint64_t
     {
     };
 
@@ -16,7 +18,7 @@ namespace lux::simulation::ecs
     // remaps identity. Generic enum serialization must not persist its bits.
     constexpr void luxBinarySemanticArchiveOnly(Entity) noexcept {}
 
-    [[nodiscard]] constexpr std::uint32_t entityBits(Entity entity) noexcept
+    [[nodiscard]] constexpr std::uint64_t entityBits(Entity entity) noexcept
     {
         return entt::to_integral(entity);
     }
